Rename MysqlStorage files to MySqlStorage
