+++ src/spider/client/Driver.cpp
@@ -14,7 +14,7 @@
 #include "../core/Error.hpp"
 #include "../core/KeyValueData.hpp"
 #include "../io/BoostAsio.hpp"  // IWYU pragma: keep
-#include "../storage/MysqlStorage.hpp"
+#include "../storage/MySqlStorage.hpp"
 #include "Exception.hpp"
 
 namespace spider {
